+++ src/ArUtil/I18N/AutoSummarize.php
@@ -128,7 +128,7 @@
  * @link      http://www.ar-php.org 
  */
 
-namespace ArUtil;
+namespace ArUtil\I18N;
 
 /**
  * This PHP class do automatic keyphrase extraction to provide a quick
