+++ src/ArUtil/I18N/CharsetD.php
@@ -71,7 +71,7 @@
  * @link      http://www.ar-php.org
  */
 
-namespace ArUtil;
+namespace ArUtil\I18N;
 
 /**
  * This PHP class detect Arabic string character set
