+++ src/ArUtil/I18N/Date.php
@@ -152,7 +152,7 @@
  * @link      http://www.ar-php.org 
  */
 
-namespace ArUtil;
+namespace ArUtil\I18N;
 
 /**
  * This PHP class is an Arabic customization for PHP date function
