+++ src/ArUtil/I18N/Glyphs.php
@@ -69,7 +69,7 @@
  * @link      http://www.ar-php.org 
  */
 
-namespace ArUtil;
+namespace ArUtil\I18N;
 
 /**
  * This PHP class render Arabic text by performs Arabic glyph joining on it
