+++ src/ArUtil/I18N/Identifier.php
@@ -88,7 +88,7 @@
  * @link      http://www.ar-php.org 
  */
 
-namespace ArUtil;
+namespace ArUtil\I18N;
 
 /**
  * This PHP class identify Arabic text segments
