+++ src/ArUtil/I18N/KeySwap.php
@@ -82,7 +82,7 @@
  * @link      http://www.ar-php.org 
  */
 
-namespace ArUtil;
+namespace ArUtil\I18N;
 
 /**
  * This PHP class convert keyboard language programmatically (English - Arabic)
