+++ src/ArUtil/I18N/Mktime.php
@@ -106,7 +106,7 @@
  * @link      http://www.ar-php.org 
  */
 
-namespace ArUtil;
+namespace ArUtil\I18N;
 
 /**
  * This PHP class is an Arabic customization for PHP mktime function
