+++ src/ArUtil/I18N/Numbers.php
@@ -101,7 +101,7 @@
  * @link      http://www.ar-php.org 
  */
 
-namespace ArUtil;
+namespace ArUtil\I18N;
 
 /**
  * This PHP class spell numbers in the Arabic idiom
