+++ src/ArUtil/I18N/Soundex.php
@@ -95,7 +95,7 @@
  * @link      http://www.ar-php.org 
  */
 
-namespace ArUtil;
+namespace ArUtil\I18N;
 
 /**
  * This PHP class implement Arabic soundex algorithm
