+++ src/ArUtil/I18N/Standard.php
@@ -62,7 +62,7 @@
  * @link      http://www.ar-php.org 
  */
 
-namespace ArUtil;
+namespace ArUtil\I18N;
 
 /**
  * This PHP class standardize Arabic text
