+++ src/ArUtil/I18N/Stemmer.php
@@ -92,7 +92,7 @@
  * @link      http://www.ar-php.org 
  */
 
-namespace ArUtil;
+namespace ArUtil\I18N;
 
 /**
  * This PHP class get stem of an Arabic word
