+++ src/ArUtil/I18N/WordTag.php
@@ -104,7 +104,7 @@
  * @link      http://www.ar-php.org 
  */
 
-namespace ArUtil;
+namespace ArUtil\I18N;
 
 /**
  * This PHP class to tagging Arabic Word
